Create BaseVRInputToEvents base class; Remove SteamVR dependencies
