+++ Scripts/Input/BaseVRInputToEvents.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ad8ea9da1084be44ca7547705f4ae91d
+timeCreated: 1487123775
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
